+++ src/main/java/subway/dao/LineDao.java
@@ -1,20 +1,70 @@
 package subway.dao;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
+import javax.sql.DataSource;
+import org.springframework.dao.support.DataAccessUtils;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.RowMapper;
+import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
+import org.springframework.stereotype.Repository;
 import subway.domain.Line;
 
-public interface LineDao {
+@Repository
+public class LineDao {
 
-    Line insert(Line line);
+    private final JdbcTemplate jdbcTemplate;
+    private final SimpleJdbcInsert insertAction;
 
-    List<Line> findAll();
+    private RowMapper<Line> rowMapper = (rs, rowNum) ->
+            new Line(
+                    rs.getLong("id"),
+                    rs.getString("name"),
+                    rs.getString("color")
+            );
 
-    Optional<Line> findById(Long id);
+    public LineDao(JdbcTemplate jdbcTemplate, DataSource dataSource) {
+        this.jdbcTemplate = jdbcTemplate;
+        this.insertAction = new SimpleJdbcInsert(dataSource)
+                .withTableName("line")
+                .usingGeneratedKeyColumns("id");
+    }
 
-    void update(Line newLine);
+    public Line insert(Line line) {
+        Map<String, Object> params = new HashMap<>();
+        params.put("id", line.getId());
+        params.put("name", line.getName());
+        params.put("color", line.getColor());
 
-    void deleteById(Long id);
+        Long lineId = insertAction.executeAndReturnKey(params).longValue();
+        return new Line(lineId, line.getName(), line.getColor());
+    }
 
-    boolean existByName(String name);
+    public List<Line> findAll() {
+        String sql = "select id, name, color from LINE";
+        return jdbcTemplate.query(sql, rowMapper);
+    }
+
+    public Optional<Line> findById(Long id) {
+        String sql = "select id, name, color from LINE WHERE id = ?";
+        return Optional.ofNullable(
+                DataAccessUtils.singleResult(jdbcTemplate.query(sql, rowMapper, id)));
+    }
+
+    public void update(Line newLine) {
+        String sql = "update LINE set name = ?, color = ? where id = ?";
+        jdbcTemplate.update(sql,
+                new Object[]{newLine.getName(), newLine.getColor(), newLine.getId()});
+    }
+
+    public void deleteById(Long id) {
+        jdbcTemplate.update("delete from Line where id = ?", id);
+    }
+
+    public boolean existByName(String name) {
+        String sql = "select count(*) from LINE where name = ? ";
+        return Boolean.TRUE.equals(jdbcTemplate.queryForObject(sql, Boolean.class, name));
+    }
 }
