--- src/main/java/subway/dao/LineDaoImpl.java
@@ -1,76 +1,0 @@
-package subway.dao;
-
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-import javax.sql.DataSource;
-import org.springframework.dao.support.DataAccessUtils;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
-import org.springframework.stereotype.Repository;
-import subway.domain.Line;
-
-@Repository
-public class LineDaoImpl implements LineDao {
-
-    private final JdbcTemplate jdbcTemplate;
-    private final SimpleJdbcInsert insertAction;
-
-    private RowMapper<Line> rowMapper = (rs, rowNum) ->
-            new Line(
-                    rs.getLong("id"),
-                    rs.getString("name"),
-                    rs.getString("color")
-            );
-
-    public LineDaoImpl(JdbcTemplate jdbcTemplate, DataSource dataSource) {
-        this.jdbcTemplate = jdbcTemplate;
-        this.insertAction = new SimpleJdbcInsert(dataSource)
-                .withTableName("line")
-                .usingGeneratedKeyColumns("id");
-    }
-
-    @Override
-    public Line insert(Line line) {
-        Map<String, Object> params = new HashMap<>();
-        params.put("id", line.getId());
-        params.put("name", line.getName());
-        params.put("color", line.getColor());
-
-        Long lineId = insertAction.executeAndReturnKey(params).longValue();
-        return new Line(lineId, line.getName(), line.getColor());
-    }
-
-    @Override
-    public List<Line> findAll() {
-        String sql = "select id, name, color from LINE";
-        return jdbcTemplate.query(sql, rowMapper);
-    }
-
-    @Override
-    public Optional<Line> findById(Long id) {
-        String sql = "select id, name, color from LINE WHERE id = ?";
-        return Optional.ofNullable(
-                DataAccessUtils.singleResult(jdbcTemplate.query(sql, rowMapper, id)));
-    }
-
-    @Override
-    public void update(Line newLine) {
-        String sql = "update LINE set name = ?, color = ? where id = ?";
-        jdbcTemplate.update(sql,
-                new Object[]{newLine.getName(), newLine.getColor(), newLine.getId()});
-    }
-
-    @Override
-    public void deleteById(Long id) {
-        jdbcTemplate.update("delete from Line where id = ?", id);
-    }
-
-    @Override
-    public boolean existByName(String name) {
-        String sql = "select count(*) from LINE where name = ? ";
-        return Boolean.TRUE.equals(jdbcTemplate.queryForObject(sql, Boolean.class, name));
-    }
-}
