--- src/main/java/subway/dao/StationDaoImpl.java
@@ -1,72 +1,0 @@
-package subway.dao;
-
-import java.util.List;
-import java.util.Optional;
-import javax.sql.DataSource;
-import org.springframework.dao.support.DataAccessUtils;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
-import org.springframework.jdbc.core.namedparam.SqlParameterSource;
-import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
-import org.springframework.stereotype.Repository;
-import subway.domain.Station;
-
-@Repository
-public class StationDaoImpl implements StationDao {
-
-    private final JdbcTemplate jdbcTemplate;
-    private final SimpleJdbcInsert insertAction;
-
-    private RowMapper<Station> rowMapper = (rs, rowNum) ->
-            new Station(
-                    rs.getLong("id"),
-                    rs.getString("name")
-            );
-
-
-    public StationDaoImpl(JdbcTemplate jdbcTemplate, DataSource dataSource) {
-        this.jdbcTemplate = jdbcTemplate;
-        this.insertAction = new SimpleJdbcInsert(dataSource)
-                .withTableName("station")
-                .usingGeneratedKeyColumns("id");
-    }
-
-    @Override
-    public Station insert(Station station) {
-        SqlParameterSource params = new BeanPropertySqlParameterSource(station);
-        Long id = insertAction.executeAndReturnKey(params).longValue();
-        return new Station(id, station.getName());
-    }
-
-    @Override
-    public List<Station> findAll() {
-        String sql = "select * from STATION";
-        return jdbcTemplate.query(sql, rowMapper);
-    }
-
-    @Override
-    public Optional<Station> findById(Long id) {
-        String sql = "select * from STATION where id = ?";
-        return Optional.ofNullable(
-                DataAccessUtils.singleResult(jdbcTemplate.query(sql, rowMapper, id)));
-    }
-
-    @Override
-    public void update(Station newStation) {
-        String sql = "update STATION set name = ? where id = ?";
-        jdbcTemplate.update(sql, new Object[]{newStation.getName(), newStation.getId()});
-    }
-
-    @Override
-    public void deleteById(Long id) {
-        String sql = "delete from STATION where id = ?";
-        jdbcTemplate.update(sql, id);
-    }
-
-    @Override
-    public boolean existByName(String name) {
-        String sql = "select count(*) from STATION where name = ? ";
-        return Boolean.TRUE.equals(jdbcTemplate.queryForObject(sql, Boolean.class, name));
-    }
-}
