+++ src/test/java/subway/dao/SectionDaoTest.java
@@ -28,18 +28,18 @@ class SectionDaoTest {
     private Section section3;
 
     @Autowired
-    private LineDaoImpl lineDaoImpl;
+    private LineDao lineDao;
 
     @Autowired
-    private StationDaoImpl stationDaoImpl;
+    private StationDao stationDao;
 
     @BeforeEach
     void setUp() {
-        Line line = lineDaoImpl.findById(1L).get();
-        Station station1 = stationDaoImpl.findById(1L).get();
-        Station station2 = stationDaoImpl.findById(2L).get();
-        Station station3 = stationDaoImpl.findById(3L).get();
-        Station station4 = stationDaoImpl.findById(4L).get();
+        Line line = lineDao.findById(1L).get();
+        Station station1 = stationDao.findById(1L).get();
+        Station station2 = stationDao.findById(2L).get();
+        Station station3 = stationDao.findById(3L).get();
+        Station station4 = stationDao.findById(4L).get();
 
         section1 = new Section(line, station1, station2, new Distance(10L));
         section2 = new Section(line, station2, station3, new Distance(10L));
